chore: moved storybooks stories directory to src
